requirements.txt
asgiref==3.6.0
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.31.0
colorama==0.4.6
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.16
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-phonenumber-field==7.0.2
django-summernote==0.8.20.0
django-view-table==1.0.1
django-zurb-foundation==5.5.0
djangorestframework==3.14.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==2.0.0.post0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
npm==0.1.1
oauthlib==3.2.2
optional-django==0.1.0
phonenumberslite==8.13.4
Pillow==9.5.0
psycopg2==2.9.5
pycountry==22.3.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.43
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.10.0
